Refactored code, moved components to separate files
